4.020 (4/21/2011)
NATPunchthroughClient: Fixed internal IP addresses incorrectly reported, sometimes causing failed LAN connections.
XBOX: Allow localization of SignIn dialog.
32 to 64 bit communications: Have stringCompressor not use unsigned int for 32/64 bit network support.	Fix error with BitStream writing SystemAddress using unsigned long, caused incorrect RakNetGUID values.
Linux (LIBCAT): Reduce entropy gathering rate to once a second.  For Linux, move vmstat -s polling to the invariant sources so it isn't run periodically.
PS3: Added Console_ShowGUIInvitationsToRooms.
PS3: Renamed Notification_Console_Got_Room_Invite_PS3 to Notification_Console_Accepted_Room_Invite_PS3
PS3: Added working implementation of  Notification_Console_Got_Room_Invite_PS3

+++ DependentExtensions/Swig/SwigWindowsCSharpSample/SwigTestApp/SwigFiles/RakNetPINVOKE.cs
@@ -1327,19 +1327,19 @@ class RakNetPINVOKE {
   public static extern void RakString_Serialize__SWIG_1(string jarg1, HandleRef jarg2);
 
   [DllImport("RakNet", EntryPoint="CSharp_RakString_SerializeCompressed__SWIG_0")]
-  public static extern void RakString_SerializeCompressed__SWIG_0(HandleRef jarg1, HandleRef jarg2, int jarg3, bool jarg4);
+  public static extern void RakString_SerializeCompressed__SWIG_0(HandleRef jarg1, HandleRef jarg2, byte jarg3, bool jarg4);
 
   [DllImport("RakNet", EntryPoint="CSharp_RakString_SerializeCompressed__SWIG_1")]
-  public static extern void RakString_SerializeCompressed__SWIG_1(HandleRef jarg1, HandleRef jarg2, int jarg3);
+  public static extern void RakString_SerializeCompressed__SWIG_1(HandleRef jarg1, HandleRef jarg2, byte jarg3);
 
   [DllImport("RakNet", EntryPoint="CSharp_RakString_SerializeCompressed__SWIG_2")]
   public static extern void RakString_SerializeCompressed__SWIG_2(HandleRef jarg1, HandleRef jarg2);
 
   [DllImport("RakNet", EntryPoint="CSharp_RakString_SerializeCompressed__SWIG_3")]
-  public static extern void RakString_SerializeCompressed__SWIG_3(string jarg1, HandleRef jarg2, int jarg3, bool jarg4);
+  public static extern void RakString_SerializeCompressed__SWIG_3(string jarg1, HandleRef jarg2, byte jarg3, bool jarg4);
 
   [DllImport("RakNet", EntryPoint="CSharp_RakString_SerializeCompressed__SWIG_4")]
-  public static extern void RakString_SerializeCompressed__SWIG_4(string jarg1, HandleRef jarg2, int jarg3);
+  public static extern void RakString_SerializeCompressed__SWIG_4(string jarg1, HandleRef jarg2, byte jarg3);
 
   [DllImport("RakNet", EntryPoint="CSharp_RakString_SerializeCompressed__SWIG_5")]
   public static extern void RakString_SerializeCompressed__SWIG_5(string jarg1, HandleRef jarg2);
